Folder editing implemented. Unpolished, but works.
Key inputs won't update on keypress, and therefore won't be validated on
the fly, because anything under forEach changed rerenders the entire
tree. What more, even observables under observableArrays can't be bound
under forEach, what use is it then?

KO is damn quirky.

Keys cannot be reordered, but can be cut and pasted to reorganize, which
is fine for now.